factory

+++ TrafficSim/src/com/company/Bus.java
@@ -2,9 +2,9 @@ package com.company;
 
 final public class Bus extends Vehicle
 {
-    private Double numberPassengers;
+    private int numberPassengers;
 
-    public Bus(MovementStatus ms, String color, Double size, Double weight, Double maxSpeed, Reputation rep, DamageStatus ds, Double np)
+    public Bus(MovementStatus ms, String color, Double size, Double weight, Double maxSpeed, Reputation rep, DamageStatus ds, int np)
     {
         super(ms, color, size, weight, maxSpeed, rep, ds);
         numberPassengers = np;
